requirements.txt
joblib
wandb
hydra-core
numpy
scipy==1.13.1
pandas
scikit-learn
pytorch-lightning
transformers
accelerate
peft
geos
reverse_geocoder
matplotlib
geoopt
einops
torchdiffeq
webdataset==0.2.57